Скорость в результирующей таблица считается вещественным делением, а не целым.

+++ trunk/kgparserSrv/src/main/java/ru/klavogonki/kgparser/processing/playersTable/PlayersResultsTable.java
@@ -135,7 +135,7 @@ public class PlayersResultsTable extends JsonObject
 				}
 			}
 
-			double averageSpeed = speedSum / playerRoundsCount;
+			double averageSpeed = ((double) speedSum) / playerRoundsCount;
 			double averageErrorPercentage = errorsPercentageSum / playerRoundsCount;
 
 			row.addCell( Integer.toString(playerRoundsCount) );
